refactor(finance): drop dead cache re-exports, point consumers at core (PRD-097)
