+++ apps/pops-api/src/modules/finance/imports/imports.e2e.test.ts
@@ -14,7 +14,7 @@ import {
   seedEntity,
   seedTransaction,
 } from '../../../shared/test-utils.js';
-import { clearCache } from './lib/ai-categorizer.js';
+import { clearCache } from '../../core/ai-usage/cache.js';
 import { transformAmex } from './transformers/amex.js';
 
 import type { Database } from 'better-sqlite3';
